requirements.txt
flwr
flwr-datasets
tensorflow-cpu
